+++ src/vgm.h
@@ -75,13 +75,13 @@ typedef struct
 } VgmCommand;
   
 
-Vgm *vgm_create (const unsigned char *buffer, int offset, size_t size);
+Vgm *vgm_create (const unsigned char *buffer, size_t offset, size_t size);
 void vgm_free (Vgm *vgm);
 size_t vgm_get_tags (const Vgm *vgm, char *dst, size_t size);
 
 uint32_t vgm_get_attr(const Vgm *vgm, int attribute);
 
-size_t vgm_next_command(const Vgm *vgm, int offset, VgmCommand *command);
+size_t vgm_next_command(const Vgm *vgm, size_t offset, VgmCommand *command);
 
 int vgm_process_command(FILE *fp, VgmCommand *command);
